requirements.txt
Flask==2.0.3
numpy==1.22.3
Pillow==9.0.1
gunicorn==20.1.0
